Add a control node
A control node which will act as an intermidiary between the motors and
the control signals.  Currently it will only accept joy messages but is
set up so that it should be easy to add an autonomous signal.  There is
also a standby mode which will send stop signals continuosly.

The control is setup so that on an xbox controller the controls are:
A: TELEOP mode
B: STANDBY mode
X: AUTONOMOUS mode
LS y-axis: Linear velocity
LS x-axis: Angular velocity