Fix Motor Controller
The motor controller has been adjusted so that the GPIOs are configured coreectly
and at the beginning of the program.  The direction is also output through the
pins now and the speed is controlled through a pwm duty cycle wired to an enable
pin.  This has been tested using LEDs and appears to work as expected.